+++ CHANGELOG.md
@@ -12,6 +12,7 @@ strict improvements (e.g. from empty strings or exceptions to something more
 reasonable).
 
 ## [Unreleased]
+- [Add published time metadata](https://github.com/mozilla/readability/pull/813)
 - [Expanded comma detection to non-Latin commas](https://github.com/mozilla/readability/pull/796)
 
 ## [0.4.4] - 2023-03-31